requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
python-multipart>=0.0.5,<0.1.0
pillow>=10.0.0
python-jose[cryptography]>=3.3.0
pyzbar>=0.1.8
pandas>=2.0.0
plotly>=5.13.0
Jinja2==2.11.3
MarkupSafe==2.0.1
aiofiles>=23.2.1  # For static file serving
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.1,<3.0.0
